create repository type in frontend and fix ImportGithubForm.tsx
