+++ frontend/src/components/Issue.ts
@@ -1,7 +1,6 @@
 
 export type Issue ={
-    number:number,
     title:string,
     created_at:string,
-    login:string
+    state:string
 }
