+++ frontend/src/components/Repository.ts
@@ -1,0 +1,6 @@
+
+
+export type Repository ={
+    created_at:string,
+    name:string
+}
